chore: localdriver should not be Go mod dependency
The localdriver repo is used in a test. But localdriver depends on this
repo, so adding it as a Go module dependency is bad because it results
in a circular dependency. The test now just gets the latest localdriver
when it runs, rather than having a fixed verison. This is not perfect,
but much better than a circular dependency.