Add nfs fixtures
- Fixtures with username/password and readonly was being skipped silently.
That's because it fails when you run it against a NFS server. for readonly
we needed to add (ro) as one of the options, but according to this commit https://github.com/cloudfoundry/nfsv3driver/blob/44424a7a3029575c855e95ca3f04d054801dc925/mapfs_mounter_test.go#L241-L259 we can't do that. username and password was also
getting omitted.

This set of fixtures is just removing those sets since it wasn't being tested